+++ doc/sphinx-guides/source/api/changelog.rst
@@ -8,6 +8,11 @@ API Changelog
 6.1
 ---
 
+New
+~~~
+- **/api/admin/clearThumbnailFailureFlag**: See :ref:`thumbnail_reset`.
+- **/api/admin/downloadTmpFile**: See :ref:`download-file-from-tmp`.
+
 Changes
 ~~~~~~~
 - **/api/datasets/{id}/versions/{versionId}/citation**: This endpoint now accepts a new boolean optional query parameter "includeDeaccessioned", which, if enabled, causes the endpoint to consider deaccessioned versions when searching for versions to obtain the citation. See :ref:`get-citation`.